In Scenario.cxx:: OldProgressPrintout Scenario.cxx:: ProgressPrintout :  added (double) before year variable ton convert to double. Otherwise, infinite loop in progress bar for scenario with total time lower than one year.

+++ source/src/Scenario.cxx
@@ -675,8 +675,8 @@ void Scenario::Evolution(cSecond t)
 //________________________________________________________________________
 void Scenario::OldProgressPrintout(cSecond t)
 {
-	double Time = (fAbsoluteTime-fStartingTime)/cYear ;
-	double Total = (t-fStartingTime)/cYear;
+	double Time = (fAbsoluteTime-fStartingTime)/(double)cYear ;
+	double Total = (t-fStartingTime)/(double)cYear;
 
 	// Reset the line
 	for(int i = 0; i < 10; i++)
@@ -701,8 +701,8 @@ void Scenario::OldProgressPrintout(cSecond t)
 //________________________________________________________________________
 void Scenario::ProgressPrintout(cSecond t)
 {
-	double Time = (fAbsoluteTime-fStartingTime)/cYear ;
-	double Total = (t-fStartingTime)/cYear;
+	double Time = (fAbsoluteTime-fStartingTime)/(double)cYear ;
+	double Total = (t-fStartingTime)/(double)cYear;
 
 	if(fOldProgressBar)
 		OldProgressPrintout(t);
